requirements.txt
boto3
cdk-ecr-deployment==2.0.2
aws-cdk-lib>=2.0.0
aws_cdk.aws_glue_alpha>=2.0.0alpha1